🚀 feat(ReactNativeFlipper.java): add release flavor of ReactNativeFlipper This commit adds the release flavor of ReactNativeFlipper, which is an empty class that does not initialize Flipper on release. This is done to prevent Flipper from being loaded on release builds, which can cause performance issues.
🔨 chore(android): update build.gradle to use latest versions of build tools and gradle
This commit updates the build.gradle file to use the latest versions of build tools and gradle. The buildToolsVersion is now set to 33.0.0, compileSdkVersion and targetSdkVersion are set to 33. The gradle version is updated to 7.3.1. The ndkVersion is set to 23.1.7779620 which has both M1 support and is the side-by-side NDK version from AGP.

🔧 chore(android): enable Hermes JS engine
This commit enables the Hermes JS engine by setting the hermesEnabled property to true in gradle.properties. This will improve the performance of the application by using a more efficient JavaScript engine.