🚀 chore(android): add com.facebook.react plugin to build.gradle This commit adds the com.facebook.react plugin to the build.gradle file in the android/app directory. This plugin is required for React Native applications and enables the use of the React Native framework in the Android application.
🔧 chore(build.gradle): clean up and organize configuration block
This commit cleans up and organizes the configuration block in the build.gradle file. The commented-out code has been removed, and the configuration block has been reorganized to make it easier to read and understand. The enableHermes variable has been moved to the react block, and the enableSeparateBuildPerCPUArchitecture and enableProguardInReleaseBuilds variables have been left unchanged.

🔧 chore(build.gradle): change applicationId namespace to com.lofftmobileapp
The applicationId namespace has been changed to com.lofftmobileapp to match the company's domain name. This change will help to avoid conflicts with other applications that may have the same applicationId.

🔥 chore(build.gradle): remove unused dependencies and exclusions
This commit removes unused dependencies and exclusions from the build.gradle file. The implementation of the fileTree function and the com.facebook.react:react-native dependency are no longer needed. The exclusion of com.facebook.fbjni and com.facebook.flipper from the debugImplementation of com.facebook.flipper:flipper is also removed.

🔥 refactor(build.gradle): remove unused code and simplify dependencies
This commit removes unused code and simplifies dependencies in the build.gradle file. The flipper-fresco-plugin dependency is now only included in debug builds. The hermes-android dependency is now used instead of the hermes-engine dependency. The isNewArchitectureEnabled function and the copyDownloadableDepsToLibs task are removed as they are no longer needed.