🔥 refactor(FlatMap.tsx): remove commented out code 👷 chore(FlatMap.tsx): use useSelector hook to get flats from Redux store The commented out code has been removed to improve code readability and maintainability. The useSelector hook has been used to get the flats from the Redux store, which improves the code by removing the need to pass the flats as a prop.
🔥 refactor(FlatFindScreen.tsx): remove unused flats prop from FlatListSubScreen and FlatMap components
✨ feat(FlatFindScreen.tsx): add dispatch to setAllFlats action creator
The flats prop was removed from the FlatListSubScreen and FlatMap components as it was not being used. The setAllFlats action creator was added to dispatch the flats data to the Redux store. This allows the flats data to be accessed from other components in the application.

🔍 refactor(FlatListSubScreen.tsx): use Redux to get flats data
The flats data is now obtained from the Redux store using the useSelector hook instead of being passed as a prop. This improves the separation of concerns and makes the component more reusable.