🔍 refactor(firestoreActions.ts): extract date format conversion to a helper function ✨ feat(firestoreActions.ts): add additional flat data fields to getFlatsFromDB function The date format conversion has been extracted to a helper function to improve readability and maintainability. Additional flat data fields have been added to the getFlatsFromDB function, including description, flatFeatures, flatMate, fromDate, untilDate, and warmRent. This provides more comprehensive data for the flats and improves the user experience.
🔧 chore(LofftHeaderPhoto.tsx): refactor LofftHeaderPhoto component to accept images as props
The LofftHeaderPhoto component now accepts an array of images as a prop, which makes it more flexible and reusable. The images are now passed as an object with a uri property instead of being imported from the assets folder. The onViewableItemsChanged function has been refactored to use the any type instead of an explicit object type.

🐛 fix(dateFormatConverter.ts): handle null date and convert seconds to date
The dateFormatConverter function now handles null dates by returning null. It also converts dates in seconds to a Date object using the convertSecondsTODate function.

🐛 fix(FlatFindScreen.tsx): remove console.log statement
The console.log statement was added for debugging purposes and is no longer needed.

🔨 refactor(FlatShowScreen.tsx): use flat description from state instead of hardcoded string
✨ feat(FlatShowScreen.tsx): add support for displaying flat images and dates
The hardcoded string for the flat description has been replaced with the description from the state. This improves maintainability and readability of the code. Support for displaying flat images and dates has been added. The LofftHeaderPhoto component now accepts an array of images to display, and the flat's available dates are now displayed dynamically.

🎨 style(FlatShowScreen.tsx): conditionally render Read More button
🚧 chore(FlatShowScreen.tsx): add comment for future development
The Read More button is now only rendered if the description length is greater than 200 characters. This improves the user experience by not showing the button when it is not needed. A comment has been added to indicate that future development is required to match the values common with the user.