🚚 chore(App.tsx): move AdminScreen to admin folder 🚚 chore(GuestNavigator.tsx): move SignUpScreen and SignInScreen to auth folder 🔧 chore(package.json): update pod-reset script to remove ios/Pods/ directory The changes in App.tsx and GuestNavigator.tsx are just moving files to a more appropriate folder structure. The pod-reset script in package.json has been updated to remove the ios/Pods/ directory as well, which is necessary for a clean install of the pods.
✨ feat(adminScreen.tsx): add AdminScreen component
This commit adds the AdminScreen component which displays a list of users and their details. The component uses the getUsersFromNotion function to fetch user data from Notion API and displays it in a list. The component also provides a sync button to update user details. The component is accessible via the navigation menu and is only available to authenticated users.

✨ feat(SignInScreen.tsx): add SignInScreen component
The SignInScreen component is added to the application. It contains a form for signing in, a button for signing in with social media accounts, and a link to the sign-up screen. The component also includes styling for the form and background images.

✨ feat(SignUpScreen.tsx): add SignUpScreen component
The SignUpScreen component is added to the application. It contains a form for users to sign up and a link to navigate to the sign-in screen. The component also includes a background image and styling to improve the user interface.