🔨 refactor(flatsSlice.ts): change allFlats state type to any[] 🔨 refactor(userImageUploadSlice.ts): change imagesToUpload state type to any[] The state types of allFlats and imagesToUpload have been changed to any[] to improve the readability of the code. The PayloadAction type has been used to specify the type of the action payload in the setAllFlats and deleteImageToUpload reducers.
🔨 refactor(userJourneySlice.ts): add types to UserJourneyState and UserJourneyActions
✨ feat(userJourneySlice.ts): add initial state to userJourneySlice
The UserJourneyState and UserJourneyActions interfaces have been added to provide type safety to the userJourneySlice. The initial state has also been added to the userJourneySlice to provide a default state for the application.

🐛 fix(userJourneySlice.ts): remove optional chaining from data.flatMate
The optional chaining operator was causing the flatMate property to be set to undefined when data.flatMate was undefined. This was causing issues with the registration process. Removing the optional chaining operator ensures that the flatMate property is set to the correct value.