🔊 chore(App.tsx): remove console.log statement 🔊 chore(firebaseAuth.ts): add console.log statement to handleSignIn function 🔊 chore(firebaseStorage.ts): add console.log statement to libraryImageUpload and uploadUserImages functions Removed a console.log statement from App.tsx. Added console.log statements to handleSignIn, libraryImageUpload, and uploadUserImages functions to improve error handling and debugging.
🔇 chore(firestoreActions.ts): remove console.log statements
This commit removes all console.log statements from the firestoreActions.ts file. These statements were used for debugging purposes and are no longer needed.

🐛 fix(firestoreActions.ts): add descriptive message to console.log in saveFlatToUserLikes function
The console.log statement in the saveFlatToUserLikes function now includes a descriptive message to make it easier to identify where the error occurred. This improves the readability of the logs and makes it easier to debug issues.

🔥 refactor(Maps/FlatMap.tsx): remove console.log statements
🔥 refactor(buttons/HeaderPageContentSwitch.tsx): remove console.log statements
The console.log statements were removed from the codebase to improve code cleanliness and remove unnecessary logging.

🔥 refactor(ListFlatApplicationCard.tsx): remove console.log statement
The console.log statement was removed from the code as it was not providing any useful information and was only cluttering the console output.

🔥 refactor(LofftHeaderPhoto.tsx): remove console.log statement
🔨 refactor(HighlightedButtons.tsx): remove unnecessary console.log statement
The console.log statements were removed from the code as they were not providing any useful information and were only cluttering the console output.

🔥 refactor(userJourneySlice.ts): remove console.log statements
🔥 refactor(usersSlice.ts): remove console.log statement
This commit removes console.log statements from the codebase. These statements are not necessary and can clutter the console output.

🔥 refactor(FlatFindScreen.tsx): remove console.log statements
The console.log statements were removed from the FlatFindScreen component. These statements were used for debugging purposes and are no longer needed.