🐛 fix(firestoreActions.ts): fix the logic of saving and removing flats to/from user likes The previous implementation of saving and removing flats to/from user likes was incorrect. The fix updates the logic to correctly add and remove flats to/from the user's savedFlats and likedUsers arrays.
🔥 refactor(ListViewFlatCard.tsx): use Redux to manage saved flats state
The state of whether a flat is saved or not is now managed by Redux. The `useState` hook has been replaced with `useAppSelector` and `useAppDispatch` hooks from the `@ReduxCore/hooks` module. The `saveFlatToUserLikes` function has been replaced with the `saveFlatsToFavorites` action creator from the `usersSlice` module. This improves the maintainability of the code and makes it easier to manage the state of saved flats across the application.

🔥 refactor(MapViewFlatCard.tsx): replace useState with useAppSelector and useAppDispatch
💾 chore(MapViewFlatCard.tsx): update saveFlatToUserLikes to saveFlatsToFavorites
The useState hook has been replaced with useAppSelector and useAppDispatch hooks from the Redux toolkit. This improves the performance of the component by reducing the number of re-renders. The saveFlatToUserLikes function has been updated to saveFlatsToFavorites to better reflect its functionality.

🔨 refactor(HighlithgtedButtons.tsx): remove unused imports and variables, rename props
✨ feat(HighlithgtedButtons.tsx): add onPressHeart and save props
This commit removes unused imports and variables, and renames the navigation prop to onPressLeftArrow. It also adds two new props, onPressHeart and save, which are used to handle the press event of the heart icon and to determine whether the heart icon should be filled or not. This improves the reusability of the component and makes it more flexible.

🚀 feat(flatsSlice.ts): add createAsyncThunk import and extraReducers field
This commit adds the createAsyncThunk import to the flatsSlice.ts file. This is necessary for creating asynchronous actions. The extraReducers field is also added to the flatsSlice object. This field is used to define reducers for actions generated by createAsyncThunk.

🚀 feat(usersSlice.ts): add saveFlatsToFavorites async thunk
🐛 fix(usersSlice.ts): initialize savedFlats as an array of strings
The saveFlatsToFavorites async thunk allows users to save or remove flats from their favorites list. The savedFlats property in the UserState interface is now initialized as an array of strings to avoid type errors.

🚀 feat(FlatShowScreen.tsx): add functionality to save flats to favorites
🐛 fix(FlatShowScreen.tsx): remove unused variable i
The saveFlatsToFavorites function is now imported from the usersSlice and dispatched when the heart icon is pressed. The save variable is used to determine whether the flat is already saved to favorites or not. The unused variable i has been removed.