🔧 chore(android): add Firebase Installation plugin to build.gradle 🐛 fix(android): change package name from com.lofftmobileapp to app.lofft.lofft 🔥 refactor(android): remove enableHermes variable and comment out its usage The Firebase Installation plugin has been added to the build.gradle file. The package name has been changed from com.lofftmobileapp to app.lofft.lofft to match the project's new name. The enableHermes variable has been removed and its usage has been commented out for bug and error testing.
🚀 chore(MainApplication.java): update package name and comment out unused package
This commit updates the package name to app.lofft.lofft to match the project's new name. It also comments out the SplashScreenReactPackage import as it is not being used and is causing issues during the upgrade to version 0.71.4.

🚀 chore(android): update android build dependencies
This commit updates the android build dependencies to use the latest version of the Google Services plugin, version 4.3.15. This is necessary for the React Native Firebase setup. Additionally, the hermesEnabled property in gradle.properties is set to true, but it may need to be set to false if errors continue to occur during testing. Finally, the package name in ReactNativeFlipper.java is updated to match the new package name of the app.