🎨 style(App.tsx): add StatusBar component to hide status bar 🐛 fix(AndroidManifest.xml): change windowSoftInputMode to adjustPan The StatusBar component is added to hide the status bar. The windowSoftInputMode is changed to adjustPan to prevent the keyboard from resizing the screen and hiding the input fields.
🎨 style(SignInWith.tsx): improve sign in with section UI
The sign in with section UI has been improved by adding horizontal lines above and below the text "Or sign in with". The text has also been centered and given some margin on the sides. The container for the sign in with section has been given a width of 100% to ensure it takes up the full width of the screen.

🔥 refactor(PrimaryScreen.tsx): remove TouchableWithoutFeedback and Keyboard.dismiss() function
The TouchableWithoutFeedback and Keyboard.dismiss() function were removed from the PrimaryScreen component. This was done to simplify the code and remove unnecessary functionality.

🎨 style(SignInScreen.tsx): add PrimaryScreen component and remove whitespace
🎨 style(CoreDesignStyleSheet.ts): add height to CoreStyleSheet
The PrimaryScreen component is added to the SignInScreen component to improve consistency with the application's design. The whitespace is removed from the SignInScreen component to improve readability. The height property is added to the CoreStyleSheet to ensure that the view takes up the full height of the screen.